Revert "improve_swagger_models"
